* moved the GenericLevel, GenericDTT and GenericPowerOnOff models to a dedicated files
