+++ bluetooth_mesh/models/generic/__init__.py
@@ -1,4 +1,4 @@
 from .onoff import *
-#from .level import *
-#from .dtt import *
-#from .ponoff import *
+from .level import *
+from .dtt import *
+from .ponoff import *
