+++ bluetooth_mesh/models/models.py
@@ -39,13 +39,6 @@ from bluetooth_mesh.messages.config import (
     SecureNetworkBeacon,
     StatusCode,
 )
-from bluetooth_mesh.messages.generic.level import GenericLevelOpcode
-from bluetooth_mesh.messages.generic.dtt import GenericDTTOpcode
-from bluetooth_mesh.messages.generic.ponoff import (
-    GenericPowerOnOffOpcode,
-    GenericPowerOnOffSetupOpcode,
-    GenericOnPowerUp,
-)
 from bluetooth_mesh.messages.health import HealthOpcode
 from bluetooth_mesh.messages.properties import PropertyID
 from bluetooth_mesh.messages.scene import SceneOpcode
@@ -79,13 +72,7 @@ __all__ = [
     "DebugClient",
     "NetworkDiagnosticClient",
     "NetworkDiagnosticSetupClient",
-    "GenericDTTServer",
-    "GenericDTTClient",
-    "GenericPowerOnOffServer",
-    "GenericPowerOnOffSetupServer",
-    "GenericPowerOnOffClient",
     "SceneClient",
-    "GenericLevelClient",
     "SensorClient",
     "GatewayConfigServer",
     "GatewayConfigClient",
@@ -1289,276 +1276,6 @@ class NetworkDiagnosticSetupClient(Model):
     PUBLISH = True
 
 
-class GenericDTTServer(Model):
-    MODEL_ID = (None, 0x1004)
-    OPCODES = {
-        GenericDTTOpcode.GENERIC_DTT_SET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericDTTClient(Model):
-    MODEL_ID = (None, 0x1005)
-    OPCODES = {
-        GenericDTTOpcode.GENERIC_DTT_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_dtt(
-        self,
-        destination: int,
-        app_index: int,
-        transition_time: float,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> float:
-        status_opcode = GenericDTTOpcode.GENERIC_DTT_STATUS
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_SET,
-                params=dict(
-                    transition_time=transition_time,
-                ),
-            )
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return status[status_opcode.name.lower()]["transition_time"]
-
-
-    async def set_dtt_unack(
-        self,
-        destination: int,
-        app_index: int,
-        transition_time: float,
-        *,
-        send_interval: float = 0.1,
-        retransmissions: int = 6,
-    ):
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_SET_UNACKNOWLEDGED,
-                params=dict(
-                    transition_time=transition_time,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def get_dtt_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[float, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericDTTOpcode.GENERIC_DTT_STATUS
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class GenericPowerOnOffServer(Model):
-    MODEL_ID = (None, 0x1006)
-    OPCODES = {
-        GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_GET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericPowerOnOffSetupServer(Model):
-    MODEL_ID = (None, 0x1007)
-    OPCODES = {
-        GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericPowerOnOffClient(Model):
-    MODEL_ID = (None, 0x1008)
-    OPCODES = {
-        GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_on_power_up(
-        self,
-        destination: int,
-        app_index: int,
-        on_power_up: GenericOnPowerUp,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ) -> int:
-
-        status_opcode = GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(on_power_up=on_power_up),
-        )
-
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET,
-                params=dict(
-                    on_power_up=on_power_up,
-                ),
-            )
-
-            return await ret
-
-        status = await self.query(
-            request, status, send_interval=send_interval, timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]["on_power_up"]
-
-
-    async def set_on_power_up_unack(
-        self,
-        destination: int,
-        app_index: int,
-        on_power_up: GenericOnPowerUp,
-        *,
-        send_interval: float = 0.1,
-        retransmissions: int = 6,
-    ):
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET_UNACKNOWLEDGED,
-                params=dict(
-                    on_power_up=on_power_up,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def get_on_power_up(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
 class SceneClient(Model):
     MODEL_ID = (None, 0x1205)
     OPCODES = {
@@ -1631,250 +1348,6 @@ class SceneClient(Model):
             node: self.expect_app(
                 node,
                 app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class GenericLevelClient(Model):
-    MODEL_ID = (None, 0x1003)
-    OPCODES = {
-        GenericLevelOpcode.GENERIC_LEVEL_STATUS,
-        GenericLevelOpcode.GENERIC_LEVEL_GET,
-        GenericLevelOpcode.GENERIC_LEVEL_SET,
-        GenericLevelOpcode.GENERIC_LEVEL_SET_UNACKNOWLEDGED,
-        GenericLevelOpcode.GENERIC_DELTA_SET,
-        GenericLevelOpcode.GENERIC_DELTA_SET_UNACKNOWLEDGED,
-        GenericLevelOpcode.GENERIC_MOVE_SET,
-        GenericLevelOpcode.GENERIC_MOVE_SET_UNACKNOWLEDGED,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_level(
-        self,
-        destination: int,
-        app_index: int,
-        level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_SET,
-                params=dict(
-                    level=level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def set_level_unack(
-        self,
-        destination: int,
-        app_index: int,
-        level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        retransmissions: int = 6,
-    ):
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_SET_UNACKNOWLEDGED,
-                params=dict(
-                    level=level,
-                    tid=tid,
-                    delay=current_delay,
-                    transition_time=transition_time,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def delta_set(
-        self,
-        destination: int,
-        app_index: int,
-        delta_level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_DELTA_SET,
-                params=dict(
-                    delta_level=delta_level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def move_set(
-        self,
-        destination: int,
-        app_index: int,
-        delta_level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_MOVE_SET,
-                params=dict(
-                    delta_level=delta_level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def get_level_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
                 destination=None,
                 opcode=status_opcode,
                 params=dict(),
